+++ otlet/__init__.py
@@ -27,5 +27,5 @@ from .types import *
 from .http import *
 from .exceptions import *
 
-__version__ = '0.2.1'
-__license__ = 'MIT'
+__version__ = "0.3.0"
+__license__ = "MIT"