+++ otlet/util.py
@@ -1,0 +1,5 @@
+from warnings import warn
+
+
+def deprecated(msg: str = ""):
+    warn(msg, DeprecationWarning, 2)